Make it possible to change where YolactNode publishes
Some users might want to configure where a yolact node publishes. The best way to achieve that is by making it possible to change the namespace both topics are in. Right now, it always is `/yolact_ros2`. That is great as the default, but it would be better suited to make it possible to configure that. It would make it possible to easily create multiple yolact nodes, for example.